backend/requirements.txt
# Core Django packages
Django>=4.2,<5.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0
django-environ>=0.10.0
django-extensions>=3.2.0

# Database
psycopg2-binary>=2.9.0
redis>=4.0.0
django-redis>=5.2.0

# Authentication & Security
djangorestframework-simplejwt>=5.2.0
cryptography>=3.4.8

# API Documentation
drf-spectacular>=0.26.0

# File handling & media
Pillow>=9.0.0
python-magic>=0.4.0

# Email & notifications
celery>=5.2.0
django-celery-beat>=2.4.0
django-celery-results>=2.4.0

# Utilities
python-dateutil>=2.8.0
pytz>=2023.0
requests>=2.28.0

# WebSocket support
channels>=4.0.0
channels-redis>=4.0.0

# Development tools
django-debug-toolbar>=4.0.0
factory-boy>=3.2.0
faker>=15.0.0

# Testing
pytest>=7.0.0
pytest-django>=4.5.0
pytest-cov>=4.0.0

# Linting & formatting
black>=22.0.0
flake8>=5.0.0
isort>=5.10.0

# Production
gunicorn>=20.0.0
whitenoise>=6.0.0